remove the column and bin_pat categories before merge
We prepend "query" and "ref" on the columns of the
availabile_bins.csv file before merging on the result.
If you don't remove these two fields first, the
result makes no sense.  For example, it will tell
you that the query category is both Fauzi and Dave's
if the mummer result had a Fauzi and a Dave bin.